+++ HISTORY.rst
@@ -3,6 +3,13 @@
 History
 -------
 
+0.1.1 (2016-09-26)
+~~~~~~~~~~~~~~~~~~
+
+* Fixed: Cache properties now allow to set cache value via ``foo = bar``
+  syntax when cache descriptor has ``as_property == True``
+
+
 0.1.0 (2015-11-26)
 ~~~~~~~~~~~~~~~~~~